fix(ww500_md): address PR #144 review findings
- demosaic line-tracker hypothesis selection: a failed hypothesis (0/1
  lines tracked) scored 0 from green_split and won unbeatably, poisoning
  the line table for the whole frame. Prefer the hypothesis that tracked
  the most lines; the green-split score only breaks ties. (The real bug
  of the four - could produce garbage hi-res frames.)
- rawdump CLI: clamp the KB arguments before the x1024 so huge/negative
  input cannot wrap the multiply past the bounds checks
- demosaic_repair_zeros: skip rows whose same-parity neighbour does not
  exist (h < 3) instead of reading out of bounds
- hm0360_md: idSaveDepth uint8_t -> uint32_t (native width, no wrap)

NOT taken: rounding the SCB_InvalidateDCache_by_Addr size to 32 - the
CMSIS implementation already extends op_size to cover the trailing
partial cache line (DCIMVAC ignores bits 4:0), so the change is a no-op.

+++ EPII_CM55M_APP_S/app/ww_projects/ww500_md/hm0360_md.c
@@ -45,7 +45,7 @@ static uint8_t mainCameraID;
 // restore from then on would pin the bus to the HM0360, silently starving
 // the main camera of its I2C commands (bench 12 Jul 2026: IMX708 stream-on
 // lost, sensor left in standby, every capture timing out).
-static uint8_t idSaveDepth = 0;
+static uint32_t idSaveDepth = 0;
 
 static bool hm0360MainCamera = false;
